Make it an error to not have 'mesh' cloned if testing is enabled
Before, it was just a warning to not have the 'meash' repo cloned if testing
was enabled.  But that just leads to a bunch of failing tests that don't make
it clear what the problem is.

If the user explicitly turns on testing and you know that tests are going to
all bomb, then it should error out instead of produce a warning.